requirements.txt
langchain 
ipykernel
langchain-core
langchain-community
python-dotenv
streamlit
langchain-groq
langsmith


